Extract and export the `array` serializer
