+++ packages/umi-serializer-beet/src/index.ts
@@ -1,3 +1,4 @@
+export * from './array';
 export * from './createBeetSerializer';
 export * from './bytes';
 export * from './errors';
