+++ packages/umi-serializer-beet/test/dataEnum.test.ts
@@ -1,5 +1,6 @@
 import test from 'ava';
 import { DataEnumToSerializerTuple } from '@metaplex-foundation/umi';
+import { array } from '../src/array';
 import { createBeetSerializer } from '../src';
 import { s, d } from './_helpers';
 import { tuple } from '../src/tuple';
@@ -35,7 +36,7 @@ const getSameSizeVariants = (): DataEnumToSerializerTuple<
   SameSizeVariants,
   SameSizeVariants
 > => {
-  const { struct, u8, u16, bool, array } = createBeetSerializer();
+  const { struct, u8, u16, bool } = createBeetSerializer();
   return [
     ['A', struct<any>([['value', u16()]])],
     [
