+++ packages/umi-serializer-data-view/src/index.ts
@@ -1,3 +1,4 @@
+export * from './array';
 export * from './createDataViewSerializer';
 export * from './bytes';
 export * from './errors';
