+++ packages/umi-serializer-data-view/test/array.test.ts
@@ -1,10 +1,11 @@
 import test from 'ava';
 import { Endian } from '@metaplex-foundation/umi';
+import { array } from '../src/array';
 import { createDataViewSerializer } from '../src';
 import { s, d } from './_helpers';
 
 test('prefixed (de)serialization', (t) => {
-  const { array, u8, string, u64 } = createDataViewSerializer();
+  const { u8, string, u64 } = createDataViewSerializer();
 
   // Empty.
   s(t, array(u8()), [], '00000000'); // 4-bytes prefix.
@@ -30,7 +31,7 @@ test('prefixed (de)serialization', (t) => {
 });
 
 test('fixed (de)serialization', (t) => {
-  const { array, u8, string, u64 } = createDataViewSerializer();
+  const { u8, string, u64 } = createDataViewSerializer();
 
   // Empty.
   s(t, array(u8(), { size: 0 }), [], '');
@@ -62,7 +63,7 @@ test('fixed (de)serialization', (t) => {
 });
 
 test('remainder (de)serialization', (t) => {
-  const { array, u8, string, u64 } = createDataViewSerializer();
+  const { u8, string, u64 } = createDataViewSerializer();
   const remainder = { size: 'remainder' } as const;
 
   // Empty.
@@ -91,7 +92,7 @@ test('remainder (de)serialization', (t) => {
 });
 
 test('description', (t) => {
-  const { array, u8, u16 } = createDataViewSerializer();
+  const { u8, u16 } = createDataViewSerializer();
 
   // Size.
   t.is(array(u8(), { size: 42 }).description, 'array(u8; 42)');
@@ -111,7 +112,7 @@ test('description', (t) => {
 });
 
 test('sizes', (t) => {
-  const { array, u8, u32, string } = createDataViewSerializer();
+  const { u8, u32, string } = createDataViewSerializer();
   t.is(array(u8()).fixedSize, null);
   t.is(array(u8()).maxSize, null);
   t.is(array(u8(), { size: u8() }).fixedSize, null);
